Asteroidz 6.6
Added score menu, scrolling feature in score menu, glowing stars on
separate canvas. Few bugs though, when organizing the list the name
doesn't change with it. When pressing enter and the focus is on the text
box for username it resets the whole game because it trys to go to a
directory.